+++ src/barretenberg/waffle/stdlib/mimc.hpp
@@ -7,12 +7,14 @@
 #include "../composer/mimc_composer.hpp"
 #include "./field/field.hpp"
 
-namespace pstd
+namespace plonk
+{
+namespace stdlib
 {
 
 field_t<waffle::MiMCComposer> mimc_hash(field_t<waffle::MiMCComposer> &input, field_t<waffle::MiMCComposer> &k_in);
 
 field_t<waffle::StandardComposer> mimc_hash(field_t<waffle::StandardComposer> input, field_t<waffle::StandardComposer> k_in);
 }
-
+}
 #endif
